requirements.txt
streamlit
langchain
google-generativeai
dotenv
langchain_groq
duckduckgo_search